src/file_lxo.c: avoid a potential integer overflow

+++ src/file_lxo.c
@@ -53,7 +53,7 @@ struct lxo_header
 static int header_check_lxo(const unsigned char *buffer, const unsigned int buffer_size, const unsigned int safe_header_only, const file_recovery_t *file_recovery, file_recovery_t *file_recovery_new)
 {
   const struct lxo_header *header=(const struct lxo_header *)buffer;
-  const uint64_t size=be32(header->size) + 8;
+  const uint64_t size=(uint64_t)be32(header->size) + 8;
   if(size < sizeof(struct lxo_header))
     return 0;
   if(buffer[8]=='L' && buffer[9]=='X' && buffer[10]=='O')
